password for destructive changes implemented

+++ controllers/categoryController.js
@@ -113,22 +113,26 @@ exports.categoryDeleteGet = async function (req, res, next) {
 
 exports.categoryDeletePost = async function (req, res, next) {
   try {
-    // Delete category from item's category field.
-    const items = await Item.find({ category: req.body.categoryId });
-    async function deleteCategory(item, categoryId) {
-      await item.category.pull({ _id: categoryId });
-      await item.save();
+    if (req.body.password === 'admin') {
+      // Delete category from item's category field.
+      const items = await Item.find({ category: req.body.categoryId });
+      async function deleteCategory(item, categoryId) {
+        await item.category.pull({ _id: categoryId });
+        await item.save();
+      }
+      const itemsDeletePromises = [];
+      items.forEach((item) =>
+        itemsDeletePromises.push(deleteCategory(item, req.body.categoryId))
+      );
+      await Promise.all(itemsDeletePromises);
+
+      // Delete category
+      await Category.findByIdAndDelete(req.body.categoryId);
+
+      res.redirect('/inventory/category');
+    } else {
+      res.redirect('/inventory/category/' + req.body.categoryId + '/delete');
     }
-    const itemsDeletePromises = [];
-    items.forEach((item) =>
-      itemsDeletePromises.push(deleteCategory(item, req.body.categoryId))
-    );
-    await Promise.all(itemsDeletePromises);
-
-    // Delete category
-    await Category.findByIdAndDelete(req.body.categoryId);
-
-    res.redirect('/inventory/category');
   } catch (error) {
     next(error);
   }
@@ -163,7 +167,7 @@ exports.categoryUpdatePost = [
         _id: req.params.id,
       });
 
-      if (!errors.isEmpty()) {
+      if (!errors.isEmpty() || req.body.password !== 'admin') {
         res.render('category_form', {
           title: 'Create New Category',
           errors: errors.errors,
